refactor: move helper functions (pose detection functions) from content.js to utils.js
